refactor: remove Serwist dependency and clean up project assets
- Removed "@serwist/next" dependency from package.json.
- Deleted unused Android launcher icon images from public directory.
- Updated tsconfig.json to remove Serwist typings and unnecessary libraries.
- Added new actions for equipment management:
  - addEquipment: Handles adding new equipment with file upload.
  - addLiters: Updates liters used for a specific equipment.
  - addSystemThumbnail: Updates the thumbnail for a specific equipment.
  - changeDate: Updates the maintenance date for coffee machines and treadmills.
  - changeFilters: Updates the filter change date for equipment with filters.
  - cleanFilters: Updates the cleaning date for equipment with filters.
  - getEquipments: Retrieves existing equipment data from JSON file.
- Improved file handling and data management in actions.
