+++ actions/changeDate.ts
@@ -1,0 +1,60 @@
+"use server";
+
+import fs from "fs";
+import path from "path";
+import { revalidatePath } from "next/cache";
+import { Category, Equipments, IEquipment } from "@/types/Equipment";
+import {
+  getCurrentDate,
+  getKeyFromKey,
+  getNextChangeDateInterval,
+} from "@/functions/functions";
+
+const changeDate = async (equipment: IEquipment, formData: FormData) => {
+  const changeDate = formData.get("change-date");
+
+  if (!changeDate) {
+    return;
+  }
+
+  // Get exists Nb liters
+  const filePath = path.resolve(process.cwd(), "data.json");
+  const jsonData = fs.readFileSync(filePath, { encoding: "utf8" });
+  const equipments = JSON.parse(jsonData) as Equipments;
+
+  // Find equipment
+  const mapEquipments = equipments.equipments.map((eqp) => {
+    if (
+      equipment.category === getKeyFromKey(Category, Category.COFFEE_MACHINE) &&
+      eqp.id === equipment.id
+    ) {
+      const lastChangedDescalDate =
+        eqp.model !== "Delonghi ECAM22.110.B"
+          ? changeDate.toString()
+          : getCurrentDate();
+      eqp.nextDescalingDate = getNextChangeDateInterval(
+        lastChangedDescalDate,
+        3
+      );
+    }
+
+    if (
+      equipment.category === getKeyFromKey(Category, Category.TREADMILL) &&
+      eqp.id === equipment.id
+    ) {
+      const lastChangedOilDate = changeDate.toString();
+      eqp.nextChangeOilDate = getNextChangeDateInterval(lastChangedOilDate, 3);
+    }
+
+    return eqp;
+  });
+
+  const data: Equipments = { equipments: [] };
+  data.equipments = mapEquipments;
+
+  // Update by writting file
+  fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
+  revalidatePath("/");
+};
+
+export default changeDate;
