+++ components/waterFilterSystem/AddLitersForm.tsx
@@ -1,44 +1,16 @@
-import fs from "fs";
-import path from "path";
-import { revalidatePath } from "next/cache";
-import { Equipments, IEquipment } from "@/types/Equipment";
+import { IEquipment } from "@/types/Equipment";
+import addLiters from "@/actions/addLiters";
 
 type AddLitersFormProps = {
   equipment: IEquipment;
 };
 
 const AddLitersForm = ({ equipment }: AddLitersFormProps) => {
-  const createLiter = async (formData: FormData) => {
-    "use server";
-
-    const nbLiters = formData.get("nb-liters");
-
-    // Get exists file
-    const filePath = path.resolve(process.cwd(), "data.json");
-    const jsonData = fs.readFileSync(filePath, { encoding: "utf8" });
-    const equipments = JSON.parse(jsonData) as Equipments;
-
-    // Map & Find equipment
-    const mapEquipments = equipments.equipments.map((eqp) => {
-      if (equipment.maxCapacityFilters && eqp.id === equipment.id) {
-        const liters = Number(eqp.litersUsed ?? 0) + Number(nbLiters);
-        eqp.litersUsed = liters;
-      }
-
-      return eqp;
-    });
-
-    const data: Equipments = { equipments: [] };
-    data.equipments = mapEquipments;
-
-    // Update by writting file
-    fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
-    revalidatePath("/");
-  };
+  const addLitersWithEquipment = addLiters.bind(null, equipment);
 
   return (
     <>
-      <form action={createLiter} className="flex gap-2">
+      <form action={addLitersWithEquipment} className="flex gap-2">
         <input
           type="number"
           name="nb-liters"
