+++ components/waterFilterSystem/ChangeFiltersForm.tsx
@@ -1,8 +1,5 @@
-import fs from "fs";
-import path from "path";
-import { Equipments, IEquipment } from "@/types/Equipment";
-import { revalidatePath } from "next/cache";
-import { getNextChangeFiltersDate } from "@/functions/functions";
+import changeFilters from "@/actions/changeFilters";
+import { IEquipment } from "@/types/Equipment";
 
 type ChangeFiltersFormProps = {
   equipment: IEquipment;
@@ -13,43 +10,11 @@ const ChangeFiltersForm = ({
   equipment,
   nextChangeFiltersDate,
 }: ChangeFiltersFormProps) => {
-  const changeFilters = async (formData: FormData) => {
-    "use server";
-
-    const changeFiltersDate = formData.get("change-filters-date");
-
-    if (!changeFiltersDate) {
-      return;
-    }
-
-    // Get exists file
-    const filePath = path.resolve(process.cwd(), "data.json");
-    const jsonData = fs.readFileSync(filePath, { encoding: "utf8" });
-    const equipments = JSON.parse(jsonData) as Equipments;
-
-    // Map & Find equipment
-    const mapEquipments = equipments.equipments.map((eqp) => {
-      if (equipment.maxCapacityFilters && eqp.id === equipment.id) {
-        const lastChangedFiltersDate = changeFiltersDate.toString();
-        eqp.nextChangeFiltersDate = getNextChangeFiltersDate(
-          lastChangedFiltersDate
-        );
-      }
-
-      return eqp;
-    });
-
-    const data: Equipments = { equipments: [] };
-    data.equipments = mapEquipments;
-
-    // Update by writting file
-    fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
-    revalidatePath("/");
-  };
+  const changeFiltersWithEquipment = changeFilters.bind(null, equipment);
 
   return (
     <>
-      <form action={changeFilters} className="flex gap-2">
+      <form action={changeFiltersWithEquipment} className="flex gap-2">
         <input
           type="hidden"
           name="change-filters-date"
