+++ components/waterFilterSystem/CleanFiltersForm.tsx
@@ -1,8 +1,5 @@
-import fs from "fs";
-import path from "path";
-import { Equipments, IEquipment } from "@/types/Equipment";
-import { revalidatePath } from "next/cache";
-import { getNextCleanFiltersDate } from "@/functions/functions";
+import { IEquipment } from "@/types/Equipment";
+import cleanFilters from "@/actions/cleanFilters";
 
 type CleanFiltersFormProps = {
   equipment: IEquipment;
@@ -13,43 +10,11 @@ const CleanFiltersForm = ({
   equipment,
   nextCleanFiltersDate,
 }: CleanFiltersFormProps) => {
-  const cleanFilters = async (formData: FormData) => {
-    "use server";
-
-    const cleanFiltersDate = formData.get("clean-filters-date");
-
-    if (!cleanFiltersDate) {
-      return;
-    }
-
-    // Get exists file
-    const filePath = path.resolve(process.cwd(), "data.json");
-    const jsonData = fs.readFileSync(filePath, { encoding: "utf8" });
-    const equipments = JSON.parse(jsonData) as Equipments;
-
-    // Map & Find equipment
-    const mapEquipments = equipments.equipments.map((eqp) => {
-      if (equipment.maxCapacityFilters && eqp.id === equipment.id) {
-        const lastCleanedFiltersDate = cleanFiltersDate.toString();
-        eqp.nextCleanFiltersDate = getNextCleanFiltersDate(
-          lastCleanedFiltersDate
-        );
-      }
-
-      return eqp;
-    });
-
-    const data: Equipments = { equipments: [] };
-    data.equipments = mapEquipments;
-
-    // Update by writting file
-    fs.writeFileSync(filePath, JSON.stringify(data, null, 2));
-    revalidatePath("/");
-  };
+  const cleanFiltersWithEquipment = cleanFilters.bind(null, equipment);
 
   return (
     <>
-      <form action={cleanFilters} className="flex gap-2">
+      <form action={cleanFiltersWithEquipment} className="flex gap-2">
         <input
           type="hidden"
           name="clean-filters-date"
