+++ next.config.ts
@@ -1,30 +1,5 @@
-// import type { NextConfig } from "next";
+import type { NextConfig } from "next";
 
-// const nextConfig: NextConfig = {
-//   /* config options here */
-// };
+const nextConfig: NextConfig = {};
 
-// export default nextConfig;
-
-import withSerwistInit from "@serwist/next";
-
-const withSerwist = withSerwistInit({
-  // Note: This is only an example. If you use Pages Router,
-  // use something else that works, such as "service-worker/index.ts".
-  swSrc: "app/sw.ts",
-  swDest: "public/sw.js",
-});
-
-export default withSerwist({
-  // Your Next.js config
-  // images: {
-  //   remotePatterns: [
-  //     {
-  //       protocol: "http",
-  //       hostname: "localhost",
-  //       port: "3000",
-  //       pathname: "/thumbnails/**",
-  //     },
-  //   ],
-  // },
-});
+export default nextConfig;
