+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@serwist/next": "^9.1.1",
     "next": "15.4.6",
     "react": "19.1.0",
     "react-dom": "19.1.0",
